moved cli to mi_graph module
