+++ mi_graph/__main__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+from mi_graph import cli
+cli()
